refactor: headernavigate => stacknavigator 코드와 통합
fix: StackNavigator props 명칭 수정
