+++ app/_common/components/StackNavigator.tsx
@@ -1,30 +1,48 @@
 "use client";
-import { useRouter } from "next/navigation";
-import { IconChevronLeft } from "@tabler/icons-react";
 
-interface StackNavigatorProps {
-  content: React.ReactNode;
+import { ComponentProps } from "react";
+import { useRouter } from "next/navigation";
+
+import { cn } from "../shadcn/utils";
+import Icon from "./Icon";
+
+interface StackNavigatorProps extends ComponentProps<"header"> {
+  element: React.ReactNode;
+  path?: string;
 }
 
-function StackNavigator({ content }: StackNavigatorProps) {
+function StackNavigator({
+  element,
+  path,
+  className,
+  ...props
+}: StackNavigatorProps) {
   const router = useRouter();
 
-  const handleClickPrevButton = () => {
+  const handleBack = () => {
     router.back();
   };
 
+  const handlePushPath = () => {
+    if (!path) return;
+
+    router.push(path);
+  };
+
   return (
-    <div>
-      <div className="m-2 flex justify-between p-2 py-6">
-        <IconChevronLeft
-          role="button"
-          className="min-w-[10%]"
-          onClick={handleClickPrevButton}
-        />
-        <h1 className="text-xl font-semibold">{content}</h1>
-        <div className="min-w-[10%]"></div>
-      </div>
-    </div>
+    <header
+      className={cn("relative p-4 text-center text-lg", className)}
+      {...props}
+    >
+      <Icon
+        id="chevron-left"
+        size={24}
+        className="absolute cursor-pointer"
+        onClick={!path ? handleBack : handlePushPath}
+      />
+      <h2>{element}</h2>
+    </header>
   );
 }
+
 export default StackNavigator;
