+++ app/my-page/edit/page.tsx
@@ -24,7 +24,7 @@ function EditPage() {
 
   return (
     <main className="container flex min-h-screen max-w-2xl flex-col gap-8 bg-gray-50 pb-8 dark:bg-gray-950">
-      <StackNavigator content={"수정"} />
+      <StackNavigator element={"수정"} />
       {data ? (
         <EditForm defaultValues={data} />
       ) : (
