+++ app/my-page/likes/page.tsx
@@ -23,7 +23,7 @@ function LikesPage() {
 
   return (
     <main className="container flex min-h-screen max-w-2xl flex-col gap-8 bg-gray-50 pb-8 dark:bg-gray-950">
-      <StackNavigator content={"내가 찔러보기한 사람"} />
+      <StackNavigator element={"내가 찔러보기한 사람"} />
       <section className="flex flex-col gap-2">
         {users
           ? users.map(el => <UserCard key={el.createdAt} data={el} />)
