+++ app/my-page/my-project/page.tsx
@@ -23,7 +23,7 @@ function MyProjectPage() {
 
   return (
     <main className="container flex min-h-screen max-w-2xl flex-col gap-8 bg-gray-50 pb-8 dark:bg-gray-950">
-      <StackNavigator content={"내가 생성한 프로젝트"} />
+      <StackNavigator element={"내가 생성한 프로젝트"} />
       {/* TODO: currentTab 기준으로 필터링 */}
       <section className="flex flex-col gap-2">
         {projects
