+++ app/my-page/page.tsx
@@ -30,7 +30,7 @@ function MyPage() {
 
   return (
     <main className="container flex min-h-screen max-w-2xl flex-col gap-8 bg-gray-50 pb-8 dark:bg-gray-950">
-      <StackNavigator content={"마이페이지"} />
+      <StackNavigator element={"마이페이지"} />
       <Profile data={userData} />
       <LikeCounter
         sendLikeCount={sendLikeCount}
