+++ app/my-page/requests/page.tsx
@@ -23,7 +23,7 @@ function ProjectRequestsPage() {
 
   return (
     <main className="container flex min-h-screen max-w-2xl flex-col gap-8 bg-gray-50 pb-8 dark:bg-gray-950">
-      <StackNavigator content={"내가 보낸 요청"} />
+      <StackNavigator element={"내가 보낸 요청"} />
       <section className="flex flex-col gap-2">
         {projects
           ? projects.map(el => <ProjectCard key={el.projectId} data={el} />)
